Create Member and MembersDatabase classes
Made a Member class to be used as blueprint, and MembersDatabase class that initialize an ArrayList of type Member called membersList.

Wrote a constructor on Member class that takes all the parameters and an addNewMember() method on MembersDatabase class to insert a new object of type Member to the membersList.